Added Clean Arch MVVM with Hilt, Retrofit, Coil, Compose and Unit Testing. Lint check refactoring.

+++ domain/src/main/java/com/lbg/domain/entity/WeatherEntity.kt
@@ -78,10 +78,10 @@ data class Current(
     val dewpointF: Double,
 
     @SerializedName("vis_km")
-    val visKM: Long,
+    val visKM: Double,
 
     @SerializedName("vis_miles")
-    val visMiles: Long,
+    val visMiles: Double,
 
     val uv: Long,
